feat: Front Master

See merge request s09-webmobile1-sub2/S09P12E201!137

Former-commit-id: c6035172006f4892df753163ecbf97af3c83dee0

+++ backend/src/main/java/com/ssafy/api/service/YumRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.Yum;
 
 @Repository
-public interface YumRepository extends JpaRepository<Yum, Integer> {
+public interface YumRepository extends JpaRepository<Yum, Long> {
 
 }
